Move logic into hook, and use hook
